Replace uses of now deleted NeuronInfo generated from NNS governance.proto with the API type (of the same name). //rs/nns/governance now builds 🤘
