+++ rs/nns/governance/src/neuron/mod.rs
@@ -1,5 +1,5 @@
 use crate::pb::v1::{
-    neuron::DissolveState, Neuron as NeuronProto, NeuronInfo, NeuronState, NeuronType,
+    neuron::DissolveState, Neuron as NeuronProto, NeuronState,
 };
 
 pub mod dissolve_state_and_age;
@@ -67,16 +67,6 @@ pub fn combine_aged_stakes(
         // most that can be lost due to rounding from the actual age, is always
         // less than 1 second, so this is not a problem.
         (x_stake_e8s + y_stake_e8s, total_age_seconds as u64)
-    }
-}
-
-impl NeuronInfo {
-    pub fn is_seed_neuron(&self) -> bool {
-        self.neuron_type == Some(NeuronType::Seed as i32)
-    }
-
-    pub fn is_ect_neuron(&self) -> bool {
-        self.neuron_type == Some(NeuronType::Ect as i32)
     }
 }
